Rename genesis protocol config
