+++ src/config/schemes.js
@@ -1,0 +1,15 @@
+/*
+* Scheme for proposing and rewarding contributions to an organization
+*/
+module.exports = {
+  // ************  SCHEME: Contribution reward  ************
+  /*
+    Scheme for proposing and rewarding contributions to an organization
+    https://github.com/daostack/arc/blob/master/contracts/universalSchemes/ContributionReward.sol
+
+    * contributionRewardSubmissionFee:      
+        - A fee that is to be paid for submitting a contribution
+        - The fee is expressed in the organization's token
+  */
+  contributionRewardSubmissionFee: 0
+}
