--- src/config/schemes/DxContributionReward.js
@@ -1,7 +1,0 @@
-/*
-* Scheme for proposing and rewarding contributions to an organization
-*/
-module.exports = {
-  // a fee (in the organization's token) that is to be paid for submitting a contribution
-  submissionFee: 0
-}
