+++ src/config/voting.js
